requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data Processing
pandas==2.0.3
numpy==1.24.3

# Google Sheets API
gspread==5.10.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0

# Machine Learning (既存環境と合わせる)
scikit-learn==1.3.0
joblib==1.3.2

# Development & Logging
python-json-logger==2.0.7